requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.5.0
lightgbm>=3.3.0
shap>=0.40.0
pyyaml>=6.0